auto redirects to encounter.
Need to fix gameState.
Perhaps hide windows that cannot be used right now? I.e. when moving, encounter window is not an option to view.
> movement and inventory are the only two permanent windows.

restriction on moving and encounters are not working. fix gameState.

extra screenState var? might need it to keep track of whether movement or encounter is in effect.